+++ CHANGELOG.rst
@@ -48,6 +48,7 @@ Changed
 - Changed UI constraints default values to pass the spec validation
 - Changed intra-switch EVC with a disabled switch or interface is not longer allowed to be created
 - Adapted ``mef_eline`` to ordered endpoints in a link. Endpoints for flow creation are compared with switch ids to overcome ordered endpoint.
+- EVCs UNI will be checked for disabled interfaces so the EVC is disabled as well.
 - ``primary_constraints.spf_attribute`` and ``secondary_constraints.spf_attribute`` will only be set in the database if they've been set in the request.
 - Changed UI spf_attribute to allow it to be ``default``, meaning an unset value